Changed AuthMethod back to interface
